Move the logic to Background to Work With the Sidebar Closed
