+++ domManager.js
@@ -445,55 +445,3 @@ export function setupQuickPinListener(moveTabToSpace, moveTabToPinned, moveTabTo
         }
     });
 }
-
-export function setupTabNavigationInSpaceListener(movToPrevTabInSpace, movToNextTabInSpace) {
-    chrome.runtime.onMessage.addListener(function(request, sender, sendResponse) {
-        if (request.command === "NextTabInSpace" || request.command === "PrevTabInSpace") {
-            console.log(`[TabNavigation] Received command: ${request.command}`, { request });
-            chrome.storage.local.get('spaces', async function(result) {
-                const spaces = result.spaces || [];
-                console.log("[TabNavigation] Loaded spaces from storage:", spaces);
-
-                const getTabSpace = (tabToFind) => {
-                    if (!tabToFind) {
-                        console.error("[TabNavigation] No tab found to navigate from.");
-                        return;
-                    }
-                    console.log("[TabNavigation] navigating from tab:", tabToFind);
-
-                    const spaceWithTempTab = spaces.find(space =>
-                        space.temporaryTabs.includes(tabToFind.id)
-                    );
-
-                    if (spaceWithTempTab) {
-                        console.log(`[TabNavigation] Tab ${tabToFind.id} is a temporary tab in space "${spaceWithTempTab.name}".`);
-                        return spaceWithTempTab;
-                    } else {
-                        const spaceWithBookmark = spaces.find(space =>
-                            space.spaceBookmarks.includes(tabToFind.id)
-                        );
-
-                        if (spaceWithBookmark) {
-                            console.log(`[TabNavigation] Tab ${tabToFind.id} is a bookmarked tab in space "${spaceWithBookmark.name}".`);
-                            return spaceWithBookmark
-                        } else {
-                            console.warn(`[TabNavigation] Tab ${tabToFind.id} not found in any space as temporary or bookmarked.`);
-                        }
-                    }
-                };
-                const foundTabs = await chrome.tabs.query({ active: true, currentWindow: true });
-                const currentTab = foundTabs.length > 0 ? foundTabs[0] : undefined;
-                const currentSpace = getTabSpace(currentTab);
-                if (currentTab === undefined || currentSpace === undefined) {
-                    return;
-                }
-
-                if (request.command === "NextTabInSpace") {
-                    await movToNextTabInSpace(currentTab.id, currentSpace);
-                } else if (request.command === "PrevTabInSpace") {
-                    await movToPrevTabInSpace(currentTab.id, currentSpace);
-                }
-            });
-        }
-    });
-}
